Add option to create AIK in TKC
Adding a new method to the Transient Key Context, allowing it to produce
attesting keys within the Owner hierarchy. These keys will be identical
to "normal" client keys, but will be restricted in their usage.

A few other changes were necessary to prevent duplication of code within
this new method. The code is mostly shared with the normal key creation
functionality.

Signed-off-by: Ionut Mihalcea <ionut.mihalcea@arm.com>